データ型更新

+++ database/migrations/2020_09_03_204153_create_ask_forms_table.php
@@ -18,7 +18,7 @@ class CreateAskFormsTable extends Migration
             //氏名
             $table->string('your_name', 20);
             //マイナンバー　12桁
-            $table->unsignedInteger('my_number',);
+            $table->bigInteger('my_number',);
             //パスワード4桁
             $table->unsignedInteger('password');
             //生年月日
